% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaHacker.R
\name{alphaHacker}
\alias{alphaHacker}
\title{alphaHacker}
\usage{
alphaHacker(data, min_items = NULL, n_out = 1)
}
\arguments{
\item{data}{Datensatz, der analysiert werden soll. Alle enthaltenen Items gehen in die Analyse mit ein.}

\item{min_items}{Definiert die Mindestanzahl Items, welche für die Berechnung von alpha verwendet werden soll.}

\item{n_out}{Definiert, wie viele Kombinationen ausgegeben werden sollen.}
}
\value{
Dataframe mit dem höchsten alpha-Wert. Beinhaltet den verwendeten Datensatz + das alpha.
}
\description{
alphaHacker
}
\examples{
data <- psych::bfi[1:10]
alphaHacker(data, min_items = 5, n_out = 3)
}
