% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biz_colors.R
\name{biz_colors}
\alias{biz_colors}
\title{biz_colors}
\usage{
biz_colors(palette = c("BIZ", "BSB", "LB"), ...)
}
\arguments{
\item{palette}{Farbpalette, die verwendet werden soll}

\item{...}{Namen der Farben, die aus der Farbpalette ausgewählt werden sollen}
}
\value{
Vektor mit ausgewählten Farben
}
\description{
biz_colors
}
