% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biz_pal.R
\name{biz_pal}
\alias{biz_pal}
\title{Funktion, die allf채llige Abstufungen zwischen den gegebenen Farben auff체llt}
\usage{
biz_pal(palette = "BIZ", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Farbpalette, die verwendet werden soll}

\item{reverse}{Boolean, der die Reihenfolge der Farbpalette umkehrt}

\item{...}{Weitere Argumente, die an colorRampPalette() weitergegeben werden}
}
\value{
Funktion, die in weiteren Funktionen verwendet wird.
}
\description{
Funktion, die allf채llige Abstufungen zwischen den gegebenen Farben auff체llt
}
