% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_formater.R
\name{date_formatter}
\alias{date_formatter}
\title{date_formatter}
\usage{
date_formatter(data, ...)
}
\arguments{
\item{data}{Datensatz, in dem die Datums-Spalten formatiert werden sollen.}

\item{...}{Spalten mit Daten, welche formatiert werden sollen. Kann im tidy-format aufgelistet werden.}
}
\value{
Datensatz mit bereinigten Datums-Spalten im ISO-Format
}
\description{
date_formatter
}
\examples{
df <- data.frame(
date1 = c("01.01.2020", "02.02.2021", "44212", "3", NA),
date2 = c("05.06.2019", "06.07.2022", "44250", "4", NA),
other_col = c("abc", "def", "ghi", "jkl", "mno")
)

df_formatted <- date_formatter(df, date1, date2)
print(df_formatted)
}
