% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodgy_merger.R
\name{dodgy_merger}
\alias{dodgy_merger}
\title{dodgy merger}
\usage{
dodgy_merger(directory = tcltk::tk_choose.dir(), ist_override = FALSE)
}
\arguments{
\item{directory}{Ordner, für den die Dateien erstellt werden sollen}

\item{ist_override}{Schaltet Warnung zum IST 5 stumm}
}
\value{
Erstellt ein PDF aus dem Word-Dokument der Mappe und kombiniert und schneidet die anderen PDF-Dateien
}
\description{
dodgy merger
}
