% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodgy_word2pdf.R
\name{dodgy_word2pdf}
\alias{dodgy_word2pdf}
\title{Word zu PDF mit PDF-XChange Tools}
\usage{
dodgy_word2pdf(inputs)
}
\arguments{
\item{inputs}{Pfad zu einer oder mehreren zu konvertierenden Dateien}
}
\description{
Konvertiert .docx/.docm zu PDF. Benötigt PDF-XChange Tools installiert.
}
\seealso{
\href{https://help.pdf-xchange.com/pdfxt10/}{PDF-Tools V10 CLI-Dokumentation}
}
