% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIntercorr.R
\name{findIntercorr}
\alias{findIntercorr}
\title{findIntercorr}
\usage{
findIntercorr(data)
}
\arguments{
\item{data}{data.frame mit 2 Spalten von Itemkombinationen.}
}
\value{
data.frame mit 2 Spalten von Itemkombinationen
}
\description{
Findet aus einem data.frame mit 2 Spalten alle einzigartigen Kombinationen,
wobei c("a", "b") als identisch zu c("b", "a") gewertet wird.
}
