% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemEasyness.R
\name{itemEasyness}
\alias{itemEasyness}
\title{itemEasyness}
\usage{
itemEasyness(data, min = 1, max = 4)
}
\arguments{
\item{data}{Datensatz als data.frame im wide-format. colnames müssen die Itemnamen sein.}

\item{min}{Kleinster möglicher Wert auf der Antwort-Skala}

\item{max}{Grösster möglicher Wert auf der Antwort-Skala}
}
\value{
Datensatz mit zwei Spalten: Itemname und Item-Einfachheit
}
\description{
itemEasyness
}
