% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemGetter.R
\name{itemGetter}
\alias{itemGetter}
\title{itemGetter}
\usage{
itemGetter(data, scale)
}
\arguments{
\item{data}{Datensatz als data.frame im wide-format. colnames müssen die Itemnamen sein.}

\item{scale}{Name der Skala, die aus dem Datensatz geholt werden soll.}
}
\value{
Vektor mit den Itemnamen der Skala
}
\description{
itemGetter
}
