% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfMerge.R
\name{pdfMergeR}
\alias{pdfMergeR}
\title{pdfMergeR}
\usage{
pdfMergeR(modul = NULL, directory = utils::choose.dir())
}
\arguments{
\item{modul}{Name des Moduls, falls abweichend vom gewählten Ordner}

\item{directory}{Ordner, für den die Dateien erstellt werden sollen}
}
\value{
Erstellt ein PDF aus dem Word-Dokument der Mappe und kombiniert und schneidet die anderen PDF-Dateien
}
\description{
pdfMergeR
}
