% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phCorr.R
\name{phCorr}
\alias{phCorr}
\title{phCorr}
\usage{
phCorr(data)
}
\arguments{
\item{data}{Datensatz als data.frame im wide-format. colnames m체ssen die Itemnamen sein.}
}
\value{
Datensatz mit zwei Spalten: Itemname und phc-Trennsch채rfe
}
\description{
Funktion zur Berechnung der par-whole-corrected Trennsch채rfe. Diese Funktion berechnet die Trennsch채rfe eines Items,
indem sie die Werte dieses Items vorher aus dem Score der Skala ausschliesst und reduziert somit den eigenen Bias.
}
