% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl_all_na.R
\name{repl_all_na}
\alias{repl_all_na}
\title{Replace all NA}
\usage{
repl_all_na(data, replace_with = 0)
}
\arguments{
\item{data}{Datensatz, in dem ALLE NA ersetzt werden sollen.}

\item{replace_with}{Wert, der statt NA eingesetzt werden soll}
}
\value{
Datensatz mit den ersetzten Werten
}
\description{
Replace all NA
}
\examples{
airquality |>
  repl_all_na(replace_with = 0) |>
  print()
}
