% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_biz.R
\name{scale_color_biz}
\alias{scale_color_biz}
\title{Color scale constructor}
\usage{
scale_color_biz(palette = "BIZ", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Farbpalette, die verwendet werden soll}

\item{discrete}{Boolean, indiziert ob color aesthetic diskret ist oder nicht}

\item{reverse}{Boolean, der die Reihenfolge der Farbpalette umkehrt}

\item{...}{Weitere Argumente, die an discrete_scale()
oder scale_color_gradientn() weitergeschickt werden}
}
\value{
Farben aus dem CI, die verwendet werden sollen (ersetzt scale_color_manual())
}
\description{
Color scale constructor
}
