% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_biz2.R
\name{scale_color_biz2}
\alias{scale_color_biz2}
\title{Color scale constructor v2}
\usage{
scale_color_biz2(palette = NULL)
}
\arguments{
\item{palette}{Farbpalette, die verwendet werden soll}
}
\value{
Farben aus dem CI, die verwendet werden sollen (ersetzt scale_color_manual())
}
\description{
Color scale constructor v2
}
