% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_biz2.R
\name{scale_fill_biz2}
\alias{scale_fill_biz2}
\title{Fill scale constructor v2}
\usage{
scale_fill_biz2(palette = NULL)
}
\arguments{
\item{palette}{Farbpalette, die verwendet werden soll}
}
\value{
Farben aus dem CI, die verwendet werden sollen (ersetzt scale_fill_manual())
}
\description{
Fill scale constructor v2
}
