% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2pdf.R
\name{word2pdf}
\alias{word2pdf}
\title{Word zu PDF mit PDF-XChange Tools}
\usage{
word2pdf(input)
}
\arguments{
\item{input}{Pfad zur .docx/.docm Datei}
}
\value{
Pfad zur erstellten PDF
}
\description{
Konvertiert .docx/.docm zu PDF. Benötigt PDF-XChange Tools installiert.
}
\seealso{
\href{https://help.pdf-xchange.com/pdfxt6/index.html?command-line-options_t.html}{PDF-Tools V6 CLI-Dokumentation}
}
