% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bizHelpeR-package.R
\docType{package}
\name{bizHelpeR-package}
\alias{bizHelpeR}
\alias{bizHelpeR-package}
\title{bizHelpeR: Helferfunktionen f端r BIZleR}
\description{
Beinhaltet die Farben der Corporate Identity der BIZ Bern f端r die Verwendung in ggplot2. Zudem eine Funktion f端r die automatische Generierung und Kombination verschiedener PDFs der Testergebnisse. Enthalten sind zudem verschiedene Helfer-Funktionen f端r Analysen des TZ.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DanStuder/bizHelpeR}
  \item Report bugs at \url{https://github.com/DanStuder/bizHelpeR/issues}
}

}
\author{
\strong{Maintainer}: Dan Studer \email{dan.studer@be.ch}

}
\keyword{internal}
