% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merger.R
\name{merger}
\alias{merger}
\title{merger}
\usage{
merger(directory = tcltk::tk_choose.dir(), ist_override = FALSE)
}
\arguments{
\item{directory}{Ordner, für den die Dateien erstellt werden sollen}
}
\value{
Erstellt ein PDF aus dem Word-Dokument der Mappe und kombiniert und schneidet die anderen PDF-Dateien
}
\description{
merger
}
