% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfMergeR.R
\name{pdfMergeR}
\alias{pdfMergeR}
\title{pdfMergeR}
\usage{
pdfMergeR(directory = utils::choose.dir(), modul = NULL)
}
\arguments{
\item{directory}{Ordner, für den die Dateien erstellt werden sollen}

\item{modul}{Name des Moduls, falls abweichend vom gewählten Ordner}
}
\value{
Erstellt ein PDF aus dem Word-Dokument der Mappe und kombiniert und schneidet die anderen PDF-Dateien
}
\description{
pdfMergeR
}
